+++ wayland-backend/src/rs/server_impl/common_poll.rs
@@ -15,7 +15,7 @@ use crate::{
 };
 
 #[cfg(any(target_os = "linux", target_os = "android"))]
-use rustix::event::epoll;
+use rustix::event::{epoll, Timespec};
 
 #[cfg(any(
     target_os = "dragonfly",
@@ -59,7 +59,7 @@ impl<D> InnerBackend<D> {
         Handle { handle: InnerHandle { state: self.state.clone() as Arc<_> } }
     }
 
-    pub fn poll_fd(&self) -> BorrowedFd {
+    pub fn poll_fd(&self) -> BorrowedFd<'_> {
         let raw_fd = self.state.lock().unwrap().poll_fd.as_raw_fd();
         // This allows the lifetime of the BorrowedFd to be tied to &self rather than the lock guard,
         // which is the real safety concern
@@ -83,15 +83,16 @@ impl<D> InnerBackend<D> {
 
         let poll_fd = self.poll_fd();
         let mut dispatched = 0;
+        let mut events = Vec::<epoll::Event>::with_capacity(32);
         loop {
-            let mut events = epoll::EventVec::with_capacity(32);
-            epoll::wait(poll_fd.as_fd(), &mut events, 0)?;
+            let buffer = rustix::buffer::spare_capacity(&mut events);
+            epoll::wait(poll_fd.as_fd(), buffer, Some(&Timespec::default()))?;
 
             if events.is_empty() {
                 break;
             }
 
-            for event in events.iter() {
+            for event in events.drain(..) {
                 let id = InnerClientId::from_u64(event.data.u64());
                 // remove the cb while we call it, to gracefully handle reentrancy
                 if let Ok(count) = self.dispatch_events_for(data, id) {
@@ -117,15 +118,16 @@ impl<D> InnerBackend<D> {
 
         let poll_fd = self.poll_fd();
         let mut dispatched = 0;
+        let mut events = Vec::<Event>::with_capacity(32);
         loop {
-            let mut events = Vec::with_capacity(32);
-            let nevents = unsafe { kevent(&poll_fd, &[], &mut events, Some(Duration::ZERO))? };
+            let buffer = rustix::buffer::spare_capacity(&mut events);
+            let nevents = unsafe { kevent(&poll_fd, &[], buffer, Some(Duration::ZERO))? };
 
             if nevents == 0 {
                 break;
             }
 
-            for event in events.iter().take(nevents) {
+            for event in events.drain(..) {
                 let id = InnerClientId::from_u64(event.udata() as u64);
                 // remove the cb while we call it, to gracefully handle reentrancy
                 if let Ok(count) = self.dispatch_events_for(data, id) {
@@ -180,13 +182,12 @@ impl<D> InnerBackend<D> {
                                 let evt = Event::new(
                                     EventFilter::Read(client.as_fd().as_raw_fd()),
                                     EventFlags::DELETE,
-                                    client_id.as_u64() as isize,
+                                    client_id.as_u64() as *mut _,
                                 );
 
-                                let mut events = Vec::new();
+                                let events: &mut [Event] = &mut [];
                                 unsafe {
-                                    kevent(&state.poll_fd, &[evt], &mut events, None)
-                                        .map(|_| ())?;
+                                    kevent(&state.poll_fd, &[evt], events, None).map(|_| ())?;
                                 }
                             }
                             return Err(e);