+++ wayland-client/src/event_queue.rs
@@ -124,7 +124,7 @@ where
     /// [`event_created_child!()`] macro is provided for overriding it.
     ///
     /// [`event_created_child!()`]: crate::event_created_child!()
-    #[cfg_attr(coverage, coverage(off))]
+    #[cfg_attr(unstable_coverage, coverage(off))]
     fn event_created_child(opcode: u16, _qhandle: &QueueHandle<State>) -> Arc<dyn ObjectData> {
         panic!(
             "Missing event_created_child specialization for event opcode {} of {}",
@@ -198,7 +198,7 @@ type QueueCallback<State> = fn(
 struct QueueEvent<State>(QueueCallback<State>, Message<ObjectId, OwnedFd>, Arc<dyn ObjectData>);
 
 impl<State> std::fmt::Debug for QueueEvent<State> {
-    #[cfg_attr(coverage, coverage(off))]
+    #[cfg_attr(unstable_coverage, coverage(off))]
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("QueueEvent").field("msg", &self.1).finish_non_exhaustive()
     }
@@ -348,7 +348,7 @@ impl<State> EventQueueInner<State> {
 }
 
 impl<State> std::fmt::Debug for EventQueue<State> {
-    #[cfg_attr(coverage, coverage(off))]
+    #[cfg_attr(unstable_coverage, coverage(off))]
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("EventQueue").field("handle", &self.handle).finish_non_exhaustive()
     }
@@ -590,7 +590,7 @@ pub struct QueueFreezeGuard<'a, State> {
 }
 
 impl<State> std::fmt::Debug for QueueHandle<State> {
-    #[cfg_attr(coverage, coverage(off))]
+    #[cfg_attr(unstable_coverage, coverage(off))]
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("QueueHandle").field("inner", &Arc::as_ptr(&self.inner)).finish()
     }
@@ -626,7 +626,7 @@ impl<State: 'static> QueueHandle<State> {
     ///
     /// This will cause the associated queue to block (or return `NotReady` to poll) until all
     /// [`QueueFreezeGuard`]s associated with the queue are dropped.
-    pub fn freeze(&self) -> QueueFreezeGuard<State> {
+    pub fn freeze(&self) -> QueueFreezeGuard<'_, State> {
         self.inner.lock().unwrap().freeze_count += 1;
         QueueFreezeGuard { qh: self }
     }
@@ -697,7 +697,7 @@ where
 }
 
 impl<I: Proxy, U: std::fmt::Debug, State> std::fmt::Debug for QueueProxyData<I, U, State> {
-    #[cfg_attr(coverage, coverage(off))]
+    #[cfg_attr(unstable_coverage, coverage(off))]
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("QueueProxyData").field("udata", &self.udata).finish()
     }