+++ wayland-sys/src/server.rs
@@ -163,7 +163,7 @@ pub fn wayland_server_option() -> Option<&'static WaylandServer> {
                 Ok(h) => return Some(h),
                 Err(::dlib::DlError::CantOpen(_)) => continue,
                 Err(::dlib::DlError::MissingSymbol(s)) => {
-                    log::error!("Found library {} cannot be used: symbol {} is missing.", ver, s);
+                    log::error!("Found library {ver} cannot be used: symbol {s} is missing.");
                     return None;
                 }
             }